Add fragments instead of activities
