+++ app/src/main/res/layout/fragment_permissions.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".PermissionsFragment">
 
     <TextView
         android:id="@+id/textViewTitle"
